Updates base squiggle component with size prop
