+++ apps/pragmatic-papers/src/blocks/SquiggleRule/Component.tsx
@@ -1,6 +1,5 @@
 import type { SquiggleRuleBlock as SquiggleRuleBlockProps } from 'src/payload-types'
 
-import { cn } from '@/utilities/ui'
 import React from 'react'
 import { Squiggle, SquiggleStatic } from '@/components/ui/squiggle'
 
@@ -11,16 +10,5 @@ type Props = {
 export const SquiggleRuleBlock: React.FC<Props> = ({ className, variant, size }) => {
   const SquiggleComponent = variant === 'static' ? SquiggleStatic : Squiggle
 
-  const sizeClasses = {
-    small: 'max-w-xs',
-    medium: 'max-w-md',
-    large: 'max-w-lg',
-    full: 'w-full',
-  }
-
-  return (
-    <div className={cn('mx-auto my-8', sizeClasses[size || 'medium'], className)}>
-      <SquiggleComponent />
-    </div>
-  )
+  return <SquiggleComponent className={className} size={size || 'medium'} />
 }
